+++ hub/exchange/src/adapter/__init__.py
@@ -3,3 +3,4 @@ from .hub_adapter       import HubAdapter       # NOQA
 from .aria_adapter      import AriaAdapter      # NOQA
 from .wemo_adapter      import WemoAdapter      # NOQA
 from .arduino_adapter   import ArduinoAdapter   # NOQA
+from .software_adapter  import SoftwareAdapter  # NOQA