+++ hub/exchange/src/device/timer_device.py
@@ -1,8 +1,18 @@
 import datetime
 import uuid
 import threading
+import logging
+from .device_type import DeviceType
+from .observable import Observable
+from .attribute import Attribute
+from .data_types import DataType
+import sched, time
 
-class TimerDevice():
+logger = logging.getLogger(__name__)
+
+import sched, time
+
+class TimerDevice(Observable):
     """
     TimerDevice calls a function at a certain configurable frequency
     TimerDevice(period, callback)
@@ -11,37 +21,62 @@ class TimerDevice():
     """
 
     def __init__(self, period):
-        super().__init__()
+        myType = self.getDeviceType()
+        super().__init__(myType)
         self.uuid = uuid.uuid4().bytes
         self.period = period
-        self.eventListeners = []
-        self.lastTime = datetime.datetime.now()
-    
+        self.runLock = threading.Lock()
+        self.stopCondition = threading.Condition(self.runLock)
+        self.running = False
+
+        def tickClosure():
+            self.tick(datetime.datetime.now())
+        self.timerThread = threading.Thread(target=tickClosure)
+
+    @staticmethod
+    def getDeviceType():
+        attributes = []
+        attributes.append(Attribute("time", DataType.String))
+        attributes.append(Attribute("period", DataType.Int))
+        return DeviceType("timer", "software", "cameron", attributes)
+
     """
     Get this device's UUID
     """
     def get_uuid(self):
         return self.uuid
 
-    """
-    Register a callback to be called when the timer fires
-    callback should take one datetime argument
-    """
-    def registerEventCallback(self, callback):
-        self.eventListeners.append(callback)
-
-    def tick(self):
+    def tickOnce(self, lastTime):
+        logger.info("TICK " + str(lastTime))
         time = datetime.datetime.now()
-        if time >= (self.lastTime + self.period):
-            self.lastTime = time
+        period = datetime.timedelta(seconds = self.period)
+
+        with self.stopCondition:
+            self.stopCondition.wait(self.period)
+
+        if not self.running:
+            return
+
+        if time >= (lastTime + period):
+            latestTime = time
             for listener in self.eventListeners:
-                listener(self.uuid, {"value" : time})
-        self.timer = threading.Timer(self.period.total_seconds(), self.tick)
-        self.timer.start()
+               listener(self.uuid, {"value" : str(time)})
+        else:
+            latestTime = lastTime
+        return latestTime
+
+    def tick(self, lastTime):
+        latestTime = self.tickOnce(lastTime)
+        if self.running:
+            self.tick(latestTime)
 
     def stop(self):
-        self.timer.cancel()
-        self.timer.join()
+        with self.runLock:
+            self.running = False
+            self.stopCondition.notify()
+        self.timerThread.join()
 
     def start(self):
-        self.tick()
+        with self.runLock:
+            self.running = True
+        self.timerThread.start()