+++ hub/exchange/test/integration/SoftwareDevices.py
@@ -2,17 +2,20 @@ from unittest   import TestCase
 from unittest import mock
 from unittest.mock import Mock
 from unittest.mock import patch
+import unittest
 
 from device     import Device
 from delegate    import  Delegate
 from ipc import Message
 from adapter.software_adapter import SoftwareAdapter
 from device.timer_device import TimerDevice
+from device import SoftwareDeviceFactory
+
 import uuid
 import time
 from datetime import timedelta
 
-class TestSoftwareAdapterIntegrations (TestCase):
+class TestSoftwareDevices (TestCase):
 
     def test_timer_should_fire_events(self):
         timer = TimerDevice(timedelta(seconds=2))
@@ -26,3 +29,7 @@ class TestSoftwareAdapterIntegrations (TestCase):
         time.sleep(7)
         timer.stop()
         self.assertEqual(mockDelegate.received.call_count, 3)
+
+    def test_software_device_controller_should_list_available_devices(self):
+        # The device list should contain at least one device type with a name of "timer"
+        self.assertTrue(SoftwareDeviceFactory.getAvailableDevices()[0].name == "timer")